crypto/create_key: split directory creation
The code sets the umask before creating the key directory. We may run
into a problem where, if we need to create more leading
components in that path, they will also inherit the restrictive umask.
This is problematic when bootstrapping a new instance, as it might
create the base data directory with permissions that prevents some
processes, e.g., nginx, from accessin it.