Replace lxml dependency with lxml_html_clean
The lxml library was only used for cleaning HTML and for generating XML in the
Piwigo plugin.

The Piwigo `lxml.etree` import for generating XML can be simply replaced by the builtin
`xml.etree.ElementTree`.

For maintenance reasons, lxml have recently split out the cleaning features into
a separate library called "lxml_clean_html", so we'll use that instead of
lxml (though it does itself depend on lxml).

We'd previously advocated installing lxml as an OS package, but this is less of
an issue with pre-built Python wheel distributions. So technically you'll still
end up with lxml installed, but it will come from PyPI.

Co-authored-by: Sandro Jäckel <sandro@c3d2.de>